Extract NDJSON direct model
